+++ src/app/admin/components/additionalSoftware/edit.controller.spec.js
@@ -1,0 +1,47 @@
+(function () {
+    'use strict';
+
+    describe('admin.EditAdditionalSoftwareController.controller', function () {
+        var vm, scope, $log, Mock;
+
+        beforeEach(function () {
+            module('chpl.mock', 'chpl.admin');
+
+            inject(function ($controller, $rootScope, _$log_, _Mock_) {
+                $log = _$log_;
+                Mock = _Mock_;
+
+                scope = $rootScope.$new();
+                vm = $controller('EditAdditionalSoftwareController', {
+                    software: 'software',
+                    $uibModalInstance: Mock.modalInstance
+                });
+                scope.$digest();
+            });
+        });
+
+        afterEach(function () {
+            if ($log.debug.logs.length > 0) {
+                //console.debug('\n Debug: ' + $log.debug.logs.join('\n Debug: '));
+            }
+        });
+
+        describe('housekeeping', function () {
+            it('should exist', function () {
+                expect(vm).toBeDefined();
+            });
+
+            it('should have a way to close the modal', function () {
+                expect(vm.cancel).toBeDefined();
+                vm.cancel();
+                expect(Mock.modalInstance.dismiss).toHaveBeenCalled();
+            });
+
+            it('should have a way to close the modal', function () {
+                expect(vm.save).toBeDefined();
+                vm.save();
+                expect(Mock.modalInstance.close).toHaveBeenCalled();
+            });
+        });
+    });
+})();